requirements.txt
flask
flask-sqlalchemy
flask-debugtoolbar
psycopg2
bootstrap-flask
readability-lxml
bert-extractive-summarizer
uwsgi
tensorflow-hub
numpy
scikit-learn
jhub-shibboleth-auth
spacy==2.1.3
transformers==2.2.2
neuralcoref
